refactor(image): source layer diff IDs from the parsed config
Image.Read already parses and holds the image config in Metadata.Config, and
trackReadProgress already sizes itself from Metadata.Config.RootFS.DiffIDs.
configDiffIDs was calling ConfigFile() a second time for the same data. That
is not free on the source this optimization targets: an OCI layout re-reads
and re-parses the config blob off disk on every call, unlike the manifest,
which it caches.

Its config-unavailable branch was also dead. readImageMetadata already
returns the ConfigFile() error, so Read fails before ever reaching it. Only
the diff-ID count check is a real fallback, which is why the tests for the
unreachable branches go away with the helper. The known-diff-ID test that
proves the layer is never asked moves to layer_metadata_test.go.

Also give Layer an unexported constructor so the known diff ID is set at
construction rather than poked in afterwards.
